refactor: Rename build.ts -> server.ts
Renamed build.ts mostly because it feels more like a more intuitive
filename
