+++ src/index.ts
@@ -1,7 +1,7 @@
 import * as dotenv from 'dotenv';
 dotenv.config();
 
-import { buildServer } from './build';
+import { buildServer } from './server';
 let PORT = process.env.PORT || 8080;
 
 let server = buildServer();
